+++ nall/priorityqueue.hpp
@@ -16,19 +16,19 @@ namespace nall {
   //O(log n) remove (dequeue)
   template<typename type_t> class priority_queue {
   public:
-    inline void tick(uint64_t ticks) {
+    inline void tick(unsigned ticks) {
       basecounter += ticks;
       while(heapsize && gte(basecounter, heap[0].counter)) callback(dequeue());
     }
 
     //counter is relative to current time (eg enqueue(64, ...) fires in 64 ticks);
-    //counter cannot exceed std::numeric_limits<uint64_t>::max() >> 1.
-    void enqueue(uint64_t counter, type_t event) {
-      uint64_t child = heapsize++;
+    //counter cannot exceed std::numeric_limits<unsigned>::max() >> 1.
+    void enqueue(unsigned counter, type_t event) {
+      unsigned child = heapsize++;
       counter += basecounter;
 
       while(child) {
-        uint64_t parent = (child - 1) >> 1;
+        unsigned parent = (child - 1) >> 1;
         if(gte(counter, heap[parent].counter)) break;
 
         heap[child].counter = heap[parent].counter;
@@ -42,11 +42,11 @@ namespace nall {
 
     type_t dequeue() {
       type_t event(heap[0].event);
-      uint64_t parent = 0;
-      uint64_t counter = heap[--heapsize].counter;
+      unsigned parent = 0;
+      unsigned counter = heap[--heapsize].counter;
 
       while(true) {
-        uint64_t child = (parent << 1) + 1;
+        unsigned child = (parent << 1) + 1;
         if(child >= heapsize) break;
         if(child + 1 < heapsize && gte(heap[child].counter, heap[child + 1].counter)) child++;
         if(gte(heap[child].counter, counter)) break;
@@ -69,13 +69,13 @@ namespace nall {
     void serialize(serializer &s) {
       s.integer(basecounter);
       s.integer(heapsize);
-      for(uint64_t n = 0; n < heapcapacity; n++) {
+      for(unsigned n = 0; n < heapcapacity; n++) {
         s.integer(heap[n].counter);
         s.integer(heap[n].event);
       }
     }
 
-    priority_queue(uint64_t size, function<void (type_t)> callback_ = &priority_queue_nocallback<type_t>)
+    priority_queue(unsigned size, function<void (type_t)> callback_ = &priority_queue_nocallback<type_t>)
     : callback(callback_) {
       heap = new heap_t[size];
       heapcapacity = size;
@@ -102,17 +102,17 @@ namespace nall {
 
   private:
     function<void (type_t)> callback;
-    uint64_t basecounter;
-    uint64_t heapsize;
-    uint64_t heapcapacity;
+    unsigned basecounter;
+    unsigned heapsize;
+    unsigned heapcapacity;
     struct heap_t {
-      uint64_t counter;
+      unsigned counter;
       type_t event;
     } *heap;
 
     //return true if x is greater than or equal to y
-    inline bool gte(uint64_t x, uint64_t y) {
-      return x - y < (std::numeric_limits<uint64_t>::max() >> 1);
+    inline bool gte(unsigned x, unsigned y) {
+      return x - y < (std::numeric_limits<unsigned>::max() >> 1);
     }
   };
 }
